Added VQueue tests
